backend/requirements.txt
# Empact backend – do not use _md JSON files in production; use backend/data/ copies.
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
requests>=2.31.0
beautifulsoup4>=4.12.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0
pydantic>=2.5.0
